LUS-5669 hsm: add hsm migration suport
Currently file migration is done by "lfs migrate". This has a couple
problems. It doesn't scale well to large files and/or large number of
files to copy. If an external tool (such as Robinhood) wants to
migrate a file, it has to spawn lfs to do the job, for every file,
which has a large overhead, or copy the file itself.

A solution to this issue is to move the copy operation into the HSM
copytool. The copytool already knows how to copy files, and Lustre has
control over these files. Migrating a file, from an application point
of view, becomes more or less the same as archiving a file. Since the
file copy is done by the copytool, this copy can be optimized in one
place.

This patch adds the ability for HSM to migrate a file within a single
Lustre filesystem. The posix copytool is enhanced to support the file
migration, and the lfs migrate command now supports that option with
--hsm and -a flags.

WC-bug-id: LU-6081
Reviewed-on: https://es-gerrit.dev.cray.com/155246
Tested-by: Elena Gryaznova <c17455@cray.com>

+++ lustre/utils/lfs.c
@@ -24,6 +24,7 @@
  * Use is subject to license terms.
  *
  * Copyright (c) 2011, 2017, Intel Corporation.
+ * Copyright 2015, 2019 Cray Inc. All rights reserved.
  */
 /*
  * This file is part of Lustre, http://www.lustre.org/
@@ -216,6 +217,7 @@ static inline int lfs_mirror_delete(int argc, char **argv)
 	SSM_CMD_COMMON("migrate  ")					\
 	"                 [--block|-b] [--non-block|-n]\n"		\
 	"                 [--non-direct|-D] [--verbose|-v]\n"		\
+	"                 [--hsm [--archive|-a NUM]]\n"			\
 	"                 FILENAME\n"
 
 #define SETDIRSTRIPE_USAGE						\
@@ -489,12 +491,13 @@ command_t cmdlist[] = {
 	{"hsm_state", lfs_hsm_state, 0, "Display the HSM information (states, "
 	 "undergoing actions) for given files.\n usage: hsm_state <file> ..."},
 	{"hsm_set", lfs_hsm_set, 0, "Set HSM user flag on specified files.\n"
-	 "usage: hsm_set [--norelease] [--noarchive] [--dirty] [--exists] "
-	 "[--archived] [--lost] [--archive-id NUM] <file> ..."},
+	 "usage: hsm_set [--norelease] [--noarchive] [--nomigrate] [--dirty] "
+	 "[--exists] [--archived] [--lost] [--archive-id <archive_id>] "
+	 "<file> ..."},
 	{"hsm_clear", lfs_hsm_clear, 0, "Clear HSM user flag on specified "
 	 "files.\n"
-	 "usage: hsm_clear [--norelease] [--noarchive] [--dirty] [--exists] "
-	 "[--archived] [--lost] <file> ..."},
+	 "usage: hsm_clear [--norelease] [--noarchive] [--nomigrate] "
+	 "[--dirty] [--exists] [--archived] [--lost] <file> ..."},
 	{"hsm_action", lfs_hsm_action, 0, "Display current HSM request for "
 	 "given files.\n" "usage: hsm_action <file> ..."},
 	{"hsm_archive", lfs_hsm_archive, 0,
@@ -528,6 +531,7 @@ command_t cmdlist[] = {
 	 "usage: migrate [--mdt-count|-c STRIPE_COUNT] [--directory|-d]\n"
 	 "               [--mdt-hash|-H HASH_TYPE]\n"
 	 "               [--mdt-index|-m START_MDT_INDEX] [--verbose|-v]\n"
+	 "		 [--hsm [--archive|-a <archive_id>]\n"
 	 "		 DIRECTORY\n"
 	 "\n"
 	 "migrate file objects from one OST layout to another\n"
@@ -623,19 +627,21 @@ static const char *error_loc = "syserror";
 static int
 migrate_open_files(const char *name, __u64 migration_flags,
 		   const struct llapi_stripe_param *param,
-		   struct llapi_layout *layout, int *fd_src, int *fd_tgt)
+		   struct llapi_layout *layout, int *fd_src, int *fd_tgt,
+		   char *volatile_file)
 {
-	int			 fd = -1;
-	int			 fdv = -1;
-	int			 rflags;
-	int			 mdt_index;
-	int                      random_value;
-	char			 parent[PATH_MAX];
-	char			 volatile_file[PATH_MAX];
-	char			*ptr;
-	int			 rc;
-	struct stat		 st;
-	struct stat		 stv;
+	int fd = -1;
+	int fdv = -1;
+	int rflags;
+	int mdt_index;
+	int random_value;
+	char parent[PATH_MAX];
+	char *volatile_name;
+	char vf[PATH_MAX];
+	char *ptr;
+	int rc;
+	struct stat st;
+	struct stat stv;
 
 	if (!param && !layout) {
 		error_loc = "layout information";
@@ -647,6 +653,14 @@ migrate_open_files(const char *name, __u64 migration_flags,
 		error_loc = "source file name";
 		return -ERANGE;
 	}
+
+	if (volatile_file == NULL)
+		volatile_file = vf;
+
+	if (migration_flags & LLAPI_MIGRATION_HSM)
+		volatile_name = ".HSM";
+	else
+		volatile_name = LUSTRE_VOLATILE_HDR;
 
 	strncpy(parent, name, sizeof(parent));
 	ptr = strrchr(parent, '/');
@@ -699,11 +713,11 @@ source_open:
 		if (rflags & O_DIRECT)
 			open_flags |= O_DIRECT;
 		random_value = random();
-		rc = snprintf(volatile_file, sizeof(volatile_file),
+		rc = snprintf(volatile_file, PATH_MAX,
 			      "%s/%s:%.4X:%.4X:fd=%.2d", parent,
-			      LUSTRE_VOLATILE_HDR, mdt_index,
+			      volatile_name, mdt_index,
 			      random_value, fd);
-		if (rc >= sizeof(volatile_file)) {
+		if (rc > strlen(volatile_file)) {
 			rc = -ENAMETOOLONG;
 			break;
 		}
@@ -731,7 +745,8 @@ source_open:
 	 * In case the MDT does not support creation of volatile files
 	 * we should try to unlink it.
 	 */
-	(void)unlink(volatile_file);
+	if (!(migration_flags & LLAPI_MIGRATION_HSM))
+		(void)unlink(volatile_file);
 
 	/*
 	 * Not-owner (root?) special case.
@@ -964,6 +979,105 @@ out_unlock:
 		rc = rc2;
 	}
 
+	return rc;
+}
+
+
+static int lfs_hsm_prepare_file(const char *file, struct lu_fid *fid,
+				dev_t *last_dev);
+
+/**
+ * Tell the HSM API to migrate a file
+ *
+ * \param name [IN] name of file to be migrated
+ * \param migration_flags [IN] flags
+ * \param param [IN] user provided param to define the migration destination
+ * \param layout [IN] layout generated from arguments to lfs
+ * \param archive_id [IN] archive group to be used to migrate data
+ *
+ * \retval 0 success
+ * \retval negative errno on error
+ */
+static int lfs_hsm_migrate(char *name, uint64_t migration_flags,
+			   struct llapi_stripe_param *param,
+			   struct llapi_layout *layout,
+			   int archive_id)
+{
+	struct hsm_user_request *hur;
+	char fullpath[PATH_MAX];
+	char volatile_file[PATH_MAX];
+	int fd_src = -1;
+	int fd_tgt = -1;
+	struct lu_fid dfid;
+	dev_t last_dev = 0;
+	int rc;
+
+	/* Allocate the request structure */
+	hur = llapi_hsm_user_request_alloc(1, FID_LEN + 1);
+	if (!hur) {
+		fprintf(stderr, "Cannot create the migrate request: %s\n",
+			strerror(ENOMEM));
+		return -ENOMEM;
+	}
+
+	rc = migrate_open_files(name, migration_flags, param, layout,
+				&fd_src, &fd_tgt, volatile_file);
+	if (rc) {
+		fprintf(stderr,
+			"Cannot create the temporary file for HSM migration %s\n",
+			strerror(-rc));
+		goto out_free;
+	}
+
+	hur->hur_request.hr_action = HUA_MIGRATE;
+	hur->hur_request.hr_archive_id = archive_id;
+
+	hur->hur_request.hr_flags = (migration_flags & LLAPI_MIGRATION_NONBLOCK) ?
+		0 : HSM_MIGRATION_BLOCKS;
+	hur->hur_request.hr_itemcount = 1;
+	hur->hur_request.hr_data_len = FID_LEN + 1;
+	hur->hur_user_item[0].hui_extent.offset = 0;
+	hur->hur_user_item[0].hui_extent.length = -1;
+	rc = lfs_hsm_prepare_file(name,
+				  &hur->hur_user_item[0].hui_fid,
+				  &last_dev);
+	if (rc)
+		goto out_free;
+
+	rc = llapi_fd2fid(fd_tgt, &dfid);
+	if (rc)
+		goto out_free;
+
+	rc = snprintf(hur_data(hur), FID_LEN + 1, DFID, PFID(&dfid));
+	if (rc < 0)
+		goto out_free;
+
+	/* Send the HSM request */
+	if (realpath(name, fullpath) == NULL) {
+		rc = -errno;
+		fprintf(stderr, "Could not find path '%s': %s\n",
+			name, strerror(-rc));
+		goto out_free;
+	}
+
+	rc = llapi_hsm_request(fullpath, hur);
+	if (rc)
+		fprintf(stderr, "Cannot send HSM request (use of %s): %s\n",
+			name, strerror(-rc));
+
+out_free:
+	if (fd_src > 0)
+		close(fd_src);
+	if (fd_tgt > 0)
+		close(fd_tgt);
+
+	free(hur);
+
+	if (!rc)
+		return 0;
+
+	/* Something has failed, and we need to unlink the target file */
+	unlink(volatile_file);
 	return rc;
 }
 
@@ -1220,7 +1334,7 @@ static int lfs_migrate(char *name, __u64 migration_flags,
 	int rc;
 
 	rc = migrate_open_files(name, migration_flags, param, layout,
-				&fd, &fdv);
+				&fd, &fdv, NULL);
 	if (rc < 0)
 		goto out;
 
@@ -1850,7 +1964,7 @@ static int mirror_extend_layout(char *name, struct llapi_layout *m_layout,
 	llapi_layout_comp_flags_set(m_layout, flags);
 	rc = migrate_open_files(name,
 			     LLAPI_MIGRATION_NONDIRECT | LLAPI_MIGRATION_MIRROR,
-			     NULL, m_layout, &fd, &fdv);
+			     NULL, m_layout, &fd, &fdv, NULL);
 	if (rc < 0)
 		goto out;
 
@@ -3352,6 +3466,7 @@ enum {
 	LFS_INHERIT_RR_OPT,
 	LFS_FIND_PERM,
 	LFS_PRINTF_OPT,
+	LFS_HSM_MIGRATE_OPT,
 };
 
 #ifndef LCME_USER_MIRROR_FLAGS
@@ -3396,6 +3511,7 @@ static int lfs_setstripe_internal(int argc, char **argv,
 	struct mirror_args		*last_mirror = NULL;
 	__u16				 mirror_id = 0;
 	char				 cmd[PATH_MAX];
+	__u32 archive_id = 0;
 	bool from_yaml = false;
 	bool from_copy = false;
 	char *template = NULL;
@@ -3409,6 +3525,7 @@ static int lfs_setstripe_internal(int argc, char **argv,
 	struct option long_opts[] = {
 /* find { .val = '0',	.name = "null",		.has_arg = no_argument }, */
 /* find	{ .val = 'A',	.name = "atime",	.has_arg = required_argument }*/
+	{ .val = 'a',	.name = "archive",	.has_arg = required_argument },
 	/* --block is only valid in migrate mode */
 	{ .val = 'b',	.name = "block",	.has_arg = no_argument },
 /* find	{ .val = 'B',	.name = "btime",	.has_arg = required_argument }*/
@@ -3460,7 +3577,8 @@ static int lfs_setstripe_internal(int argc, char **argv,
 /* find	{ .val = 'F',	.name = "fid",		.has_arg = no_argument }, */
 /* find	{ .val = 'g',	.name = "gid",		.has_arg = no_argument }, */
 /* find	{ .val = 'G',	.name = "group",	.has_arg = required_argument }*/
-	{ .val = 'h',	.name = "help",		.has_arg = no_argument },
+	{ .val = LFS_HSM_MIGRATE_OPT,
+			.name = "hsm",		.has_arg = no_argument },
 	{ .val = 'H',	.name = "mdt-hash",	.has_arg = required_argument},
 	{ .val = 'i',	.name = "stripe-index",	.has_arg = required_argument},
 	{ .val = 'i',	.name = "stripe_index",	.has_arg = required_argument},
@@ -3468,9 +3586,7 @@ static int lfs_setstripe_internal(int argc, char **argv,
 	{ .val = 'I',	.name = "component-id",	.has_arg = required_argument},
 /* find { .val = 'l',	.name = "lazy",		.has_arg = no_argument }, */
 	{ .val = 'L',	.name = "layout",	.has_arg = required_argument },
-	{ .val = 'm',	.name = "mdt",		.has_arg = required_argument},
 	{ .val = 'm',	.name = "mdt-index",	.has_arg = required_argument},
-	{ .val = 'm',	.name = "mdt_index",	.has_arg = required_argument},
 	/* --non-block is only valid in migrate mode */
 	{ .val = 'n',	.name = "non-block",	.has_arg = no_argument },
 	{ .val = 'N',	.name = "mirror-count",	.has_arg = optional_argument},
@@ -3511,12 +3627,20 @@ static int lfs_setstripe_internal(int argc, char **argv,
 	snprintf(cmd, sizeof(cmd), "%s %s", progname, argv[0]);
 	progname = cmd;
 	while ((c = getopt_long(argc, argv,
-				"bc:C:dDE:f:hH:i:I:m:N::no:p:L:s:S:vx:y:z:",
+				"a:bc:C:dDE:f:hH:i:I:m:N::no:p:L:s:S:vx:y:z:",
 				long_opts, NULL)) >= 0) {
 		size_units = 1;
 		switch (c) {
 		case 0:
 			/* Long options. */
+			break;
+		case 'a':
+			if (!migrate_mode) {
+				fprintf(stderr,
+					"archive number is only valid with migrate\n");
+				return CMD_HELP;
+			}
+			archive_id = atoi(optarg);
 			break;
 		case LFS_COMP_ADD_OPT:
 			comp_add = 1;
@@ -3737,6 +3861,14 @@ static int lfs_setstripe_internal(int argc, char **argv,
 					goto usage_error;
 				}
 			}
+			break;
+		case LFS_HSM_MIGRATE_OPT:
+			if (!migrate_mode) {
+				fprintf(stderr,
+					"--hsm is only valid with migrate\n\n");
+				return CMD_HELP;
+			}
+			migration_flags |= LLAPI_MIGRATION_HSM;
 			break;
 		case 'H':
 			if (!migrate_mode) {
@@ -4044,6 +4176,13 @@ static int lfs_setstripe_internal(int argc, char **argv,
 		goto error;
 	}
 
+	if (archive_id != 0 && !(migration_flags & LLAPI_MIGRATION_HSM)) {
+		fprintf(stderr,
+			"error: %s: --archive / -a has been specified without --hsm\n",
+			argv[0]);
+		return CMD_HELP;
+	}
+
 	if (mirror_mode) {
 		if (!setstripe_args_specified(&lsa))
 			last_mirror->m_inherit = true;
@@ -4317,8 +4456,14 @@ static int lfs_setstripe_internal(int argc, char **argv,
 		if (migrate_mdt_mode) {
 			result = llapi_migrate_mdt(fname, &migrate_mdt_param);
 		} else if (migrate_mode) {
-			result = lfs_migrate(fname, migration_flags, param,
-					     layout);
+			if (migration_flags & LLAPI_MIGRATION_HSM)
+				result = lfs_hsm_migrate(fname,
+							 migration_flags,
+							 param, layout,
+							 archive_id);
+			else
+				result = lfs_migrate(fname, migration_flags,
+						     param, layout);
 		} else if (comp_set != 0) {
 			result = lfs_component_set(fname, comp_id,
 						   lsa.lsa_pool_name,
@@ -9584,6 +9729,8 @@ static int lfs_hsm_state(int argc, char **argv)
 			printf(" never_release");
 		if (hus.hus_states & HS_NOARCHIVE)
 			printf(" never_archive");
+		if (hus.hus_states & HS_NOMIGRATE)
+			printf(" never_migrate");
 		if (hus.hus_states & HS_LOST)
 			printf(" lost_from_hsm");
 
@@ -9615,6 +9762,7 @@ static int lfs_hsm_change_flags(int argc, char **argv, int mode)
 	{ .val = 'h',	.name = "help",		.has_arg = no_argument },
 	{ .val = 'i',	.name = "archive-id",	.has_arg = required_argument },
 	{ .val = 'l',	.name = "lost",		.has_arg = no_argument },
+	{ .val = 'm',	.name = "nomigrate",	.has_arg = no_argument },
 	{ .val = 'r',	.name = "norelease",	.has_arg = no_argument },
 	{ .name = NULL } };
 	__u64 mask = 0;
@@ -9656,6 +9804,9 @@ static int lfs_hsm_change_flags(int argc, char **argv, int mode)
 					progname, end);
 				return CMD_HELP;
 			}
+			break;
+		case 'm':
+			mask |= HS_NOMIGRATE;
 			break;
 		default:
 			fprintf(stderr, "%s: unrecognized option '%s'\n",
@@ -9727,7 +9878,9 @@ static int lfs_hsm_action(int argc, char **argv)
 		printf(" %s ", hsm_progress_state2name(hps));
 
 		if ((hps == HPS_RUNNING) &&
-		    (hua == HUA_ARCHIVE || hua == HUA_RESTORE))
+		    (hua == HUA_ARCHIVE ||
+		     hua == HUA_RESTORE ||
+		     hua == HUA_MIGRATE))
 			printf("(%llu bytes moved)\n",
 			       (unsigned long long)he.length);
 		else if ((he.offset + he.length) == LUSTRE_EOF)
@@ -9769,9 +9922,10 @@ static int lfs_hsm_prepare_file(const char *file, struct lu_fid *fid,
 	int		rc;
 
 	rc = lstat(file, &st);
-	if (rc) {
-		fprintf(stderr, "Cannot stat %s: %s\n", file, strerror(errno));
-		return -errno;
+	if (rc == -1) {
+		rc = -errno;
+		fprintf(stderr, "Cannot stat %s: %s\n", file, strerror(-rc));
+		return rc;
 	}
 	/*
 	 * Checking for regular file as archiving as posix copytool
@@ -9917,7 +10071,7 @@ static int lfs_hsm_request(int argc, char **argv, int action)
 	if (!hur) {
 		fprintf(stderr, "Cannot create the request: %s\n",
 			strerror(errno));
-		return errno;
+		return -ENOMEM;
 	}
 	nbfile_alloc = nbfile;
